cron: Use docker healthchecks
The current example of cron is a bit over complicated. We spin up an
additional container, with all its resource cost. Need to share data
between them, all just to run a command, ideally within the same
instance.

Instead, we can happily make use of healthcheck, which is a standard
docker feature.

This serves two purposes. One, it executes the cron job as desired.
Secondly, it gives us a slightly better sense of the containers health.
E.g. if the cron-job cannot run, the container surely is in a bad health
state.

The alternative to health monitoring would be to query an endpoint
(such as ocs/v2.php/apps/serverinfo/api/v1/info) and parse that, but one
doesn't exclude the other. It is still easily possible to do both.

E.g. 'wget api-info | grep -q "<statuscode>200</statuscode>" && cron'.
Though I would expect that cron only works with a status code of 200
anyway.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>